Age/Gender/Emotion classification does not work on Linux/OSX (#208)

+++ src/AgeClassification/dlib/dnn/loss/multiclass_log/age/Age.h
@@ -9,8 +9,8 @@
 #include "defines.h"
 #include "DlibDotNet.Native.Dnn/dlib/dnn/loss/multiclass_log/template.h"
 
-typedef unsigned long age_out_type;
-typedef unsigned long age_train_label_type;
+typedef uint32_t age_out_type;
+typedef uint32_t age_train_label_type;
 
 MAKE_LOSSMULTICLASSLOG_FUNC(age_train_type, matrix_element_type::RgbPixel, dlib::rgb_pixel, matrix_element_type::UInt32, age_train_label_type, 200)
